Student Educational Career Recommendation System using ML
This project is meant to be a part of the progress of Education towards better course recommendation. It uses machine learning program that asks the client questions, and recommends the better stream based on the skills and academic performance provided. The program also serves as a data collection platform to support the drive for more data on course recommendation.